move email activity
