+++ lib/sendgrid4r/rest/email_activity.rb
