+++ spec/rest/email_activity_spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require File.dirname(__FILE__) + '/../../spec_helper'
+require File.dirname(__FILE__) + '/../spec_helper'
 
 describe SendGrid4r::REST::EmailActivity do
   describe 'integration test', :it do
